# docs/bruno/fixtures-game-api/apispec/fixtures-game-api.openapi.yaml
...
openapi: 3.0.3
info:
  title: Fixtures Game Backend API
  version: 1.0.0
  description: API contract for the no-money football prediction game backend. Use
    `FE - ...` tags for frontend work and `Internal - ...` tags for backend jobs/admin
    maintenance.
servers:
- url: http://localhost:3001
  description: Local backend
...
paths:
...
  "/friends-groups/{friendsGroupId}/request/status":
    get:
      tags:
      - FE - Groups And Invites
      summary: My join request status
      operationId: get_friends_groups_friendsGroupId_request_status
      description: 'FE placement: private invite/group pending screen. Purpose: shows
        whether the current user request is pending, approved, or rejected.'
      x-fe-scope: frontend
      x-bruno-folder: FE App APIs - Use These In Frontend/Groups And Invites
      parameters:
      - name: friendsGroupId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/JoinRequestStatusResponse"
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden for the current user.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
      - cookieAuth: []
  "/friends-groups/requests/{joinRequestId}/approve":
    post:
      tags:
      - FE - Groups And Invites
      summary: 'Owner: approve join request'
      operationId: post_friends_groups_requests_joinRequestId_approve
      description: 'FE placement: owner approve button. Purpose: approves a private
        join request and adds the user as group member.'
      x-fe-scope: frontend
      x-bruno-folder: FE App APIs - Use These In Frontend/Groups And Invites
      parameters:
      - name: joinRequestId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/JoinRequestReviewResponse"
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden for the current user.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
      - cookieAuth: []
  "/friends-groups/{friendsGroupId}/requests/pending":
    get:
      tags:
      - FE - Groups And Invites
      summary: 'Owner: list pending join requests'
      operationId: get_friends_groups_friendsGroupId_requests_pending
      description: 'FE placement: owner pending-requests badge/list. Purpose: focused
        queue of pending private join requests.'
      x-fe-scope: frontend
      x-bruno-folder: FE App APIs - Use These In Frontend/Groups And Invites
      parameters:
      - name: friendsGroupId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PendingJoinRequestsResponse"
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden for the current user.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
      - cookieAuth: []
  "/friends-groups/requests/{joinRequestId}/reject":
    post:
      tags:
      - FE - Groups And Invites
      summary: 'Owner: reject join request'
      operationId: post_friends_groups_requests_joinRequestId_reject
      description: 'FE placement: owner reject button. Purpose: rejects a private
        join request without adding membership.'
      x-fe-scope: frontend
      x-bruno-folder: FE App APIs - Use These In Frontend/Groups And Invites
      parameters:
      - name: joinRequestId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/JoinRequestReviewResponse"
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '403':
          description: Forbidden for the current user.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Resource not found.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
      security:
      - cookieAuth: []
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase session cookies set by /auth/sign-in or /auth/sign-up.
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true
      required:
      - error
...
    JoinRequestStatusResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            status:
              type: string
              enum:
              - pending
              - approved
              - rejected
          required:
          - status
          nullable: true
        message:
          type: string
    JoinRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        friends_group_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - pending
          - approved
          - rejected
        message:
          type: string
          nullable: true
        user_display_name:
          type: string
          nullable: true
        requested_at:
          type: string
          format: date-time
        processed_at:
          type: string
          format: date-time
          nullable: true
        processed_by:
          type: string
          format: uuid
          nullable: true
      required:
      - id
      - friends_group_id
      - user_id
      - status
    JoinRequestReviewResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            friends_group_id:
              type: string
              format: uuid
            user_id:
              type: string
              format: uuid
            status:
              type: string
              enum:
              - approved
              - rejected
          required:
          - id
          - friends_group_id
          - user_id
          - status
      required:
      - data
    PendingJoinRequestsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            "$ref": "#/components/schemas/JoinRequest"
      required:
      - data
...